requirements.txt
pandas
psycopg2-binary  # For PostgreSQL connection
sqlalchemy